% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotColoredMatrix.R
\name{makeGrid}
\alias{makeGrid}
\title{create grid with grid package}
\usage{
makeGrid(length, ncol)
}
\arguments{
\item{length}{\itemize{
\item the title
}}

\item{ncol}{\itemize{
\item position
}}
}
\description{
create grid with grid package
}
