% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotColoredMatrix.R
\name{plotColoredMatrix}
\alias{plotColoredMatrix}
\title{print colored matrix}
\usage{
plotColoredMatrix(vec, options = list())
}
\arguments{
\item{vec}{\itemize{
\item a vector with colors
}}

\item{options}{\itemize{
\item options
}}
}
\description{
print colored matrix
}
