% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFoldChange.R
\name{plotFoldChange}
\alias{plotFoldChange}
\title{print lipoprotein  quantification report}
\usage{
plotFoldChange(folds, options = list())
}
\arguments{
\item{options}{\itemize{
\item a list of options
}}

\item{lip}{\itemize{
\item a lipoprotein report
}}
}
\value{
print a report
}
\description{
print lipoprotein  quantification report
}
