% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusion-package.R
\docType{package}
\name{show-package}
\alias{show}
\alias{show-package}
\title{show: show: plot library for metabolic profiling}
\description{
functions to display molecular phenotyping data
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jwist/show}
  \item Report bugs at \url{https://github.com/jwist/show/issues}
}

}
\author{
\strong{Maintainer}: Julien Wist \email{julien.wist@correounivalle.edu.co} (\href{https://orcid.org/0000-0002-3416-2572}{ORCID})

}
\keyword{internal}
